third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-requireusermediation-basics.html

<!DOCTYPE html>
<title>Credential Manager: requireUserMediation() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
function stubResolverChecker(c) {
    assert_equals(c, undefined, "FIXME: We're currently always returning 'undefined'.");
    this.done();
}

function stubRejectionChecker(reason) {
    assert_unreached("requireUserMediation() should not reject, but did: " + reason.name);
}

async_test(function () {
    navigator.credentials.requireUserMediation().then(
        this.step_func(stubResolverChecker.bind(this)),
        this.step_func(stubRejectionChecker.bind(this)));
}, "Verify the basics of requireUserMediation()");
</script>
